break functions into separate files
move weekly, recent, readycheck to their own files

update to discordjs v14 gateway intents
