+++ package.json
@@ -23,5 +23,6 @@
     "dotenv": "^16.0.1",
     "request": "^2.88.2",
     "tweetnacl": "^1.0.3"
-  }
+  },
+  "devDependencies": {}
 }
