+++ players.json
@@ -90,5 +90,17 @@
     "993021386396282961" : {
         "name": "Poppy Milk",
         "id": "177186395"
+    },
+    "105819872780419072" : {
+        "name": "Mercury Sulfide",
+        "id": "82444103"
+    },
+    "514663058493276162" : {
+        "name": "tlace",
+        "id": "237910815"
+    },
+    "207539633117069313" : {
+        "name": "satman",
+        "id": "91504727"
     }
 }
